refactor: move to correct util file
